run ffmpeg with subprocess module (fix problem)
You need to pass the options to the subprocess.run function
separately in a list, you cannot merge several options into one
string. subprocess will try to break them up, which can mess up
filepaths with spaces in it, as string characters such as ''
are not considered when breaking up the option string.