use custom ArgumentParser to not quit on bad arguments
The default 'ArgumentParser' from the argparse module quits
when wrong arguments are passed. However, we do not want to quit
our server on bad arguments, we want to return and wait for the
next ones. Therefore we create a custom 'CustomArgumentParser'
that just raises a generic exception when an error occurs.
The rest of the code stays the same.