remove in_filename, as we have in_filepath already
in_filepath contains the name anyways and we set the path
when we create the cut and do not change it afterwards.
for the output we need to store both the name and the folder as
we can change both at will (from the gui) and also before starting
the cut (we make names unique,...).